move datauri as a shared package
