+++ httpapi/httpapi_test.go
@@ -13,8 +13,8 @@ import (
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
 	"github.com/textileio/bidbot/lib/auction"
+	"github.com/textileio/bidbot/lib/datauri"
 	"github.com/textileio/bidbot/lib/marketpeer"
-	"github.com/textileio/bidbot/service/datauri"
 	bidstore "github.com/textileio/bidbot/service/store"
 	golog "github.com/textileio/go-log/v2"
 )
