+++ lib/datauri/datauri.go
