+++ service/store/store.go
@@ -22,8 +22,8 @@ import (
 	"github.com/libp2p/go-libp2p-core/peer"
 	"github.com/oklog/ulid/v2"
 	"github.com/textileio/bidbot/lib/auction"
+	"github.com/textileio/bidbot/lib/datauri"
 	"github.com/textileio/bidbot/lib/dshelper/txndswrap"
-	"github.com/textileio/bidbot/service/datauri"
 	"github.com/textileio/bidbot/service/limiter"
 	"github.com/textileio/bidbot/service/lotusclient"
 	dsextensions "github.com/textileio/go-datastore-extensions"
