+++ service/store/store_test.go
@@ -19,10 +19,10 @@ import (
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
 	"github.com/textileio/bidbot/lib/auction"
+	"github.com/textileio/bidbot/lib/datauri/apitest"
 	"github.com/textileio/bidbot/lib/logging"
 	"github.com/textileio/bidbot/lib/marketpeer"
 	lotusclientmocks "github.com/textileio/bidbot/lib/mocks/lotusclient"
-	"github.com/textileio/bidbot/service/datauri/apitest"
 	"github.com/textileio/bidbot/service/limiter"
 	badger "github.com/textileio/go-ds-badger3"
 	golog "github.com/textileio/go-log/v2"
